# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "get": {
        "operationId": "UsersController_list",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Поиск по login, displayName или email",
            "schema": {
              "example": "nord",
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "USER",
                "MANAGER",
                "ADMIN"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "DISABLED"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUsersResponseDto"
                }
              }
            }
          }
        },
        "summary": "Список пользователей",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "summary": "Создать пользователя",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}": {
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "summary": "Обновить пользователя",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/reset-password": {
      "post": {
        "operationId": "UsersController_resetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetUserPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "summary": "Сбросить пароль пользователя",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/deactivate": {
      "post": {
        "operationId": "UsersController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "summary": "Деактивировать пользователя",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/activate": {
      "post": {
        "operationId": "UsersController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "summary": "Активировать пользователя",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/topics": {
      "get": {
        "operationId": "TopicsController_list",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Поиск по name или slug",
            "schema": {
              "example": "react",
              "type": "string"
            }
          },
          {
            "name": "usedOnly",
            "required": false,
            "in": "query",
            "description": "Вернуть только темы, связанные хотя бы с одним вопросом",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListTopicsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Список тем",
        "tags": [
          "topics"
        ]
      },
      "post": {
        "operationId": "TopicsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTopicDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicDto"
                }
              }
            }
          }
        },
        "summary": "Создать тему",
        "tags": [
          "topics"
        ]
      }
    },
    "/api/topics/{id}": {
      "patch": {
        "operationId": "TopicsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTopicDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicDto"
                }
              }
            }
          }
        },
        "summary": "Переименовать тему",
        "tags": [
          "topics"
        ]
      }
    },
...
    "/api/training/presets": {
      "get": {
        "operationId": "TrainingController_listPresets",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingPresetDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Список тренировочных пресетов",
        "tags": [
          "training"
        ]
      },
      "post": {
        "operationId": "TrainingController_createPreset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrainingPresetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingPresetDto"
                }
              }
            }
          }
        },
        "summary": "Создать тренировочный пресет",
        "tags": [
          "training"
        ]
      }
    },
...
    "/api/training/presets/{id}": {
      "patch": {
        "operationId": "TrainingController_updatePreset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTrainingPresetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingPresetDto"
                }
              }
            }
          }
        },
        "summary": "Обновить тренировочный пресет",
        "tags": [
          "training"
        ]
      },
      "delete": {
        "operationId": "TrainingController_removePreset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Удалить тренировочный пресет",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/prepare": {
      "post": {
        "operationId": "TrainingController_prepareTraining",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepareTrainingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrepareTrainingResponseDto"
                }
              }
            }
          }
        },
        "summary": "Подготовить тренировку по темам",
        "tags": [
          "training"
        ]
      }
    },
...
    "/api/training/history/{id}": {
      "get": {
        "operationId": "TrainingController_getHistoryDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingHistoryDetailResponseDto"
                }
              }
            }
          }
        },
        "summary": "Детали сохранённой тренировки текущего пользователя",
        "tags": [
          "training"
        ]
      }
    },
    "/api/training/results": {
      "post": {
        "operationId": "TrainingController_saveTrainingResults",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveTrainingResultsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveTrainingResultsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Сохранить результаты тренировки",
        "tags": [
          "training"
        ]
      }
    }
  },
  "info": {
    "title": "Nord API",
    "description": "Backend API for interview questions bank",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm8q4x7r10001s2t3u4v5w6x7"
          },
          "login": {
            "type": "string",
            "example": "nord.admin"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "nord.admin@example.com"
          },
          "displayName": {
            "type": "string",
            "example": "Nord Admin"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "MANAGER",
              "ADMIN"
            ],
            "example": "ADMIN"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED"
            ],
            "example": "ACTIVE"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "login",
          "email",
          "displayName",
          "role",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ListUsersResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDto"
            }
          },
          "total": {
            "type": "number",
            "example": 2
          },
          "meta": {
            "$ref": "#/components/schemas/ListUsersMetaDto"
          }
        },
        "required": [
          "items",
          "total",
          "meta"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "example": "nord.user"
          },
          "password": {
            "type": "string",
            "example": "user-password-2026",
            "minLength": 12,
            "maxLength": 128
          },
          "displayName": {
            "type": "string",
            "example": "Nord User"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "nord.user@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "MANAGER",
              "ADMIN"
            ],
            "example": "USER"
          }
        },
        "required": [
          "login",
          "password",
          "displayName",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "example": "Nord User"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "nord.user@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "MANAGER",
              "ADMIN"
            ],
            "example": "USER"
          }
        }
      },
      "ResetUserPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "example": "new-password-2026",
            "minLength": 12,
            "maxLength": 128
          }
        },
        "required": [
          "password"
        ]
      },
...
      "TopicDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm7y5z8qv0003x4f2w7sn1abc"
          },
          "name": {
            "type": "string",
            "example": "JavaScript"
          },
          "slug": {
            "type": "string",
            "example": "javascript"
          },
          "questionsCount": {
            "type": "number",
            "example": 42
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "questionsCount"
        ]
      },
...
      "ListTopicsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopicDto"
            }
          },
          "total": {
            "type": "number",
            "example": 25
          },
          "meta": {
            "$ref": "#/components/schemas/ListTopicsMetaDto"
          }
        },
        "required": [
          "items",
          "total",
          "meta"
        ]
      },
      "CreateTopicDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название темы",
            "maxLength": 100,
            "example": "Vue"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateTopicDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Новое название темы",
            "maxLength": 100,
            "example": "React Router"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "TrainingPresetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm7y5z8qv0003x4f2w7sn1preset"
          },
          "name": {
            "type": "string",
            "example": "Angular Developer"
          },
          "topics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionTopicDto"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "topics",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateTrainingPresetDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Angular Developer"
          },
          "topicIds": {
            "example": [
              "seed-topic-javascript",
              "seed-topic-typescript",
              "seed-topic-angular"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "topicIds"
        ]
      },
      "UpdateTrainingPresetDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Angular Developer"
          },
          "topicIds": {
            "example": [
              "seed-topic-javascript",
              "seed-topic-typescript",
              "seed-topic-angular"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "topicIds"
        ]
      },
      "PrepareTrainingDto": {
        "type": "object",
        "properties": {
          "topicIds": {
            "example": [
              "seed-topic-javascript",
              "seed-topic-typescript",
              "seed-topic-angular",
              "seed-topic-scss",
              "seed-topic-ngrx"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "topicIds"
        ]
      },
...
      "PrepareTrainingResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrainingQuestionItemDto"
            }
          },
          "total": {
            "type": "number",
            "example": 19
          },
          "meta": {
            "$ref": "#/components/schemas/PrepareTrainingMetaDto"
          }
        },
        "required": [
          "items",
          "total",
          "meta"
        ]
      },
...
      "TrainingHistoryDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm7y5z8qv0003x4f2w7sn1session"
          },
          "status": {
            "type": "string",
            "enum": [
              "COMPLETED",
              "ABANDONED_SAVED"
            ],
            "example": "COMPLETED"
          },
          "resultsCount": {
            "type": "number",
            "example": 12
          },
          "correctCount": {
            "type": "number",
            "example": 5
          },
          "incorrectCount": {
            "type": "number",
            "example": 4
          },
          "partialCount": {
            "type": "number",
            "example": 3
          },
          "feedback": {
            "type": "object",
            "nullable": true,
            "example": "Нужно точнее формулировать компромиссы и аргументацию."
          },
          "trainer": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/TrainingHistoryTrainerDto"
              }
            ]
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrainingHistoryResultItemDto"
            }
          }
        },
        "required": [
          "id",
          "status",
          "resultsCount",
          "correctCount",
          "incorrectCount",
          "partialCount",
          "feedback",
          "trainer",
          "finishedAt",
          "results"
        ]
      },
...
      "SaveTrainingResultsDto": {
        "type": "object",
        "properties": {
          "targetUserId": {
            "type": "string",
            "nullable": true,
            "example": "cm8q4x7r10001s2t3u4v5w6x7"
          },
          "feedback": {
            "type": "string",
            "nullable": true,
            "example": "Нужно подтянуть аргументацию по Angular change detection и RxJS error handling."
          },
          "status": {
            "type": "string",
            "enum": [
              "COMPLETED",
              "ABANDONED_SAVED"
            ],
            "example": "COMPLETED"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SaveTrainingResultsItemDto"
            }
          }
        },
        "required": [
          "status",
          "items"
        ]
      },
      "SaveTrainingResultsResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm7y5z8qv0003x4f2w7sn1session"
          },
          "status": {
            "type": "string",
            "enum": [
              "COMPLETED",
              "ABANDONED_SAVED"
            ],
            "example": "COMPLETED"
          },
          "resultsCount": {
            "type": "number",
            "example": 12
          },
          "correctCount": {
            "type": "number",
            "example": 7
          },
          "incorrectCount": {
            "type": "number",
            "example": 5
          },
          "partialCount": {
            "type": "number",
            "example": 2
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "status",
          "resultsCount",
          "correctCount",
          "incorrectCount",
          "partialCount",
          "finishedAt"
        ]
      }
    }
  }
}